apps/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
ortools==9.8.3296
openai==1.10.0
pydantic==2.6.0
pytest==8.0.0
httpx==0.26.0
reportlab==4.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.1
sqlalchemy==2.0.25
